Use robocopy on Windows
When running the cluster in a Windows shell (cmd or Powershell), cp is
an alias to copy and doesn't support POSIX-style command line
parameters (`cp -a`). We need to use the Windows alternative for
recursively copying file trees. Hopefully it also preserve anything
that needs to be preserved.

```
2021-04-09 13:25.24         worker [INFO] Exec(cp .git): "cp" "-a"
                                          "C:\Users\Antonin\var\lib\ocluster-scheduler\git\opam-repository-mingw.git-a8bab3e29df0d516af7e9b295055fd81d447940b\.git"
                                          "C:\Users\Antonin\var\lib\ocluster-scheduler\tmp\build-context-29eeaa\.git"
2021-04-09 13:25.31         worker [WARNING] Build failed: Unix.Unix_error(Unix.ENOENT, "CreateProcess", "")
2021-04-09 13:25.31      capnp-rpc [WARNING] Uncaught exception: Unix.Unix_error(Unix.ENOENT, "CreateProcess", "")
```

ENOENT is raised because `-a` as a file name doesn't exist.